Decouple content verification gate from workspace install
