+++ .github/scripts/content-verification/agentic-gate-cli.test.ts
@@ -1,6 +1,6 @@
 import assert from "node:assert/strict";
 import { execFile } from "node:child_process";
-import { mkdtemp, mkdir, rm, writeFile } from "node:fs/promises";
+import { copyFile, mkdtemp, mkdir, rm, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { describe, it } from "node:test";
@@ -17,7 +17,10 @@ const manifest = {
 };
 const script = new URL("./agentic-gate-cli.ts", import.meta.url);
 
-async function runGate(items: unknown[]): Promise<void> {
+async function runGate(
+  items: unknown[],
+  gateScript = script.pathname,
+): Promise<void> {
   const directory = await mkdtemp(join(tmpdir(), "content-verification-gate-"));
   const nested = join(directory, "nested");
   await mkdir(nested);
@@ -32,7 +35,7 @@ async function runGate(items: unknown[]): Promise<void> {
         JSON.stringify({ errors: [], items }),
       ),
     ]);
-    await executeFile(process.execPath, [script.pathname], {
+    await executeFile(process.execPath, [gateScript], {
       env: {
         ...process.env,
         CONTENT_VERIFICATION_AGENT_RESULT: "success",
@@ -46,10 +49,41 @@ async function runGate(items: unknown[]): Promise<void> {
   }
 }
 
+async function runIsolatedGate(items: unknown[]): Promise<void> {
+  const directory = await mkdtemp(
+    join(tmpdir(), "content-verification-gate-runtime-"),
+  );
+  try {
+    await Promise.all([
+      copyFile(
+        new URL("./agentic-gate-cli.ts", import.meta.url),
+        join(directory, "agentic-gate-cli.ts"),
+      ),
+      copyFile(
+        new URL("./agentic-gate.ts", import.meta.url),
+        join(directory, "agentic-gate.ts"),
+      ),
+      copyFile(
+        new URL("./scope.ts", import.meta.url),
+        join(directory, "scope.ts"),
+      ),
+    ]);
+    await runGate(items, join(directory, "agentic-gate-cli.ts"));
+  } finally {
+    await rm(directory, { force: true, recursive: true });
+  }
+}
+
 describe("content verification gate CLI", () => {
   it("accepts a downloaded noop artifact", async () => {
     await assert.doesNotReject(() =>
       runGate([{ type: "noop", message: "No issue is needed." }]),
+    );
+  });
+
+  it("runs without installed workspace dependencies", async () => {
+    await assert.doesNotReject(() =>
+      runIsolatedGate([{ type: "noop", message: "No issue is needed." }]),
     );
   });
