+++ .github/scripts/content-verification/targets.ts
@@ -6,14 +6,7 @@ import {
   type KnowledgeType,
   validateKnowledgeIndex,
 } from "@knowledge-base/knowledge-index";
-
-export const verificationScopes = [
-  "time-sensitive-knowledge",
-  "evergreen-knowledge",
-  "maintained-agent-content",
-] as const;
-
-export type VerificationScope = (typeof verificationScopes)[number];
+import type { VerificationScope } from "./scope.ts";
 
 export type VerificationTarget = {
   files: string[];
@@ -210,13 +203,4 @@ export function discoverVerificationTargets(
     }
   }
   return targets;
-}
-
-export function parseVerificationScope(value: string): VerificationScope {
-  if (!verificationScopes.includes(value as VerificationScope)) {
-    throw new Error(
-      `Verification scope must be one of: ${verificationScopes.join(", ")}.`,
-    );
-  }
-  return value as VerificationScope;
 }
